pyproject.toml
[project]
name = "ride-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "markdown>=3.10.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core", "src/ride", "src/user"]
